Move create tests in mutations
