+++ packages/graphql/tests/api-v6/integration/mutations/create/types/array/number-array.int.test.ts
@@ -17,8 +17,8 @@
  * limitations under the License.
  */
 
-import type { UniqueType } from "../../../../../utils/graphql-types";
-import { TestHelper } from "../../../../../utils/tests-helper";
+import type { UniqueType } from "../../../../../../utils/graphql-types";
+import { TestHelper } from "../../../../../../utils/tests-helper";
 
 describe("Create Nodes with Numeric array fields", () => {
     const testHelper = new TestHelper({ v6Api: true });
