+++ packages/graphql/tests/api-v6/integration/mutations/create/types/cartesian-point/cartesian-point-2d.int.test.ts
@@ -17,15 +17,16 @@
  * limitations under the License.
  */
 
-import type { UniqueType } from "../../../../../utils/graphql-types";
-import { TestHelper } from "../../../../../utils/tests-helper";
+import type { UniqueType } from "../../../../../../utils/graphql-types";
+import { TestHelper } from "../../../../../../utils/tests-helper";
 
-describe("Create Nodes with Point 3d", () => {
+describe("Create Nodes with CartesianPoint 2d", () => {
     const testHelper = new TestHelper({ v6Api: true });
 
     let Location: UniqueType;
-    const London = { longitude: -0.127758, latitude: 51.507351, height: 24 } as const;
-    const Rome = { longitude: 12.496365, latitude: 41.902782, height: 35 } as const;
+
+    const London = { x: -14221.955504767046, y: 6711533.711877272 } as const;
+    const Rome = { x: 1391088.9885668862, y: 5146427.7652232265 } as const;
 
     beforeEach(async () => {
         Location = testHelper.createUniqueType("Location");
@@ -33,7 +34,7 @@ describe("Create Nodes with Point 3d", () => {
         const typeDefs = /* GraphQL */ `
             type ${Location} @node {
                 id: ID!
-                value: Point!
+                value: CartesianPoint!
             }
         `;
 
@@ -44,29 +45,28 @@ describe("Create Nodes with Point 3d", () => {
         await testHelper.close();
     });
 
-    test("should create nodes with wgs-84-3d point fields", async () => {
+    test("should create nodes with wgs-84-2d point fields", async () => {
         const mutation = /* GraphQL */ `
-            mutation {
-                ${Location.operations.create}(input: [
-                        { node: { id: "1", value: { longitude: ${London.longitude}, latitude: ${London.latitude}, height: ${London.height} } } }
-                        { node: { id: "2", value: { longitude: ${Rome.longitude}, latitude: ${Rome.latitude}, height: ${Rome.height} } } }
-                    ]) 
-                    {
-                        ${Location.plural} {
-                            id
-                            value {
-                                latitude
-                                longitude
-                                height
-                                crs
-                                srid
-                            }
+        mutation {
+            ${Location.operations.create}(input: [
+                    { node: { id: "1", value: { x: ${London.x}, y: ${London.y} } }  }
+                    { node: { id: "2", value: { x: ${Rome.x}, y: ${Rome.y} } } }
+                ]) 
+                {
+                    ${Location.plural} {
+                        id
+                        value {
+                            x
+                            y
+                            crs
+                            srid
                         }
                     }
-                   
-                
-            }
-        `;
+                }
+               
+            
+        }
+    `;
 
         const mutationResult = await testHelper.executeGraphQL(mutation);
         expect(mutationResult.errors).toBeFalsy();
@@ -76,21 +76,19 @@ describe("Create Nodes with Point 3d", () => {
                     {
                         id: "1",
                         value: {
-                            latitude: London.latitude,
-                            longitude: London.longitude,
-                            height: London.height,
-                            crs: "wgs-84-3d",
-                            srid: 4979,
+                            y: London.y,
+                            x: London.x,
+                            crs: "cartesian",
+                            srid: 7203,
                         },
                     },
                     {
                         id: "2",
                         value: {
-                            latitude: Rome.latitude,
-                            longitude: Rome.longitude,
-                            height: Rome.height,
-                            crs: "wgs-84-3d",
-                            srid: 4979,
+                            y: Rome.y,
+                            x: Rome.x,
+                            crs: "cartesian",
+                            srid: 7203,
                         },
                     },
                 ]),
