+++ packages/graphql/tests/api-v6/integration/mutations/create/types/dateTime.int.test.ts
@@ -17,11 +17,10 @@
  * limitations under the License.
  */
 
-import neo4jDriver from "neo4j-driver";
-import type { UniqueType } from "../../../../utils/graphql-types";
-import { TestHelper } from "../../../../utils/tests-helper";
+import type { UniqueType } from "../../../../../utils/graphql-types";
+import { TestHelper } from "../../../../../utils/tests-helper";
 
-describe("Create Nodes with Duration fields", () => {
+describe("Create Nodes with DateTime fields", () => {
     const testHelper = new TestHelper({ v6Api: true });
     let Movie: UniqueType;
 
@@ -29,7 +28,7 @@ describe("Create Nodes with Duration fields", () => {
         Movie = testHelper.createUniqueType("Movie");
         const typeDefs = /* GraphQL */ `
         type ${Movie.name} @node {
-            duration: Duration
+            datetime: DateTime
         }
     `;
         await testHelper.initNeo4jGraphQL({ typeDefs });
@@ -38,18 +37,19 @@ describe("Create Nodes with Duration fields", () => {
     afterEach(async () => {
         await testHelper.close();
     });
-    test("should be able to create nodes with Duration fields", async () => {
-        const duration1 = new neo4jDriver.Duration(1, 2, 3, 4);
-        const duration2 = new neo4jDriver.Duration(5, 6, 7, 8);
+
+    test("should be able to create nodes with DateTime fields", async () => {
+        const date1 = new Date(1716904582368);
+        const date2 = new Date(1796904582368);
 
         const mutation = /* GraphQL */ `
             mutation {
                 ${Movie.operations.create}(input: [
-                        { node: { duration: "${duration1.toString()}" } }
-                        { node: { duration: "${duration2.toString()}" } }
+                        { node: { datetime: "${date1.toISOString()}" } }
+                        { node: { datetime: "${date2.toISOString()}" } }
                     ]) {
                     ${Movie.plural} {
-                        duration
+                        datetime
                     }
                 }
             }
@@ -61,8 +61,8 @@ describe("Create Nodes with Duration fields", () => {
         expect(gqlResult.data).toEqual({
             [Movie.operations.create]: {
                 [Movie.plural]: expect.toIncludeSameMembers([
-                    { duration: duration1.toString() },
-                    { duration: duration2.toString() },
+                    { datetime: date1.toISOString() },
+                    { datetime: date2.toISOString() },
                 ]),
             },
         });
