+++ packages/graphql/tests/api-v6/integration/mutations/create/types/duration.int.test.ts
@@ -1,0 +1,70 @@
+/*
+ * Copyright (c) "Neo4j"
+ * Neo4j Sweden AB [http://neo4j.com]
+ *
+ * This file is part of Neo4j.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import neo4jDriver from "neo4j-driver";
+import type { UniqueType } from "../../../../../utils/graphql-types";
+import { TestHelper } from "../../../../../utils/tests-helper";
+
+describe("Create Nodes with Duration fields", () => {
+    const testHelper = new TestHelper({ v6Api: true });
+    let Movie: UniqueType;
+
+    beforeEach(async () => {
+        Movie = testHelper.createUniqueType("Movie");
+        const typeDefs = /* GraphQL */ `
+        type ${Movie.name} @node {
+            duration: Duration
+        }
+    `;
+        await testHelper.initNeo4jGraphQL({ typeDefs });
+    });
+
+    afterEach(async () => {
+        await testHelper.close();
+    });
+    test("should be able to create nodes with Duration fields", async () => {
+        const duration1 = new neo4jDriver.Duration(1, 2, 3, 4);
+        const duration2 = new neo4jDriver.Duration(5, 6, 7, 8);
+
+        const mutation = /* GraphQL */ `
+            mutation {
+                ${Movie.operations.create}(input: [
+                        { node: { duration: "${duration1.toString()}" } }
+                        { node: { duration: "${duration2.toString()}" } }
+                    ]) {
+                    ${Movie.plural} {
+                        duration
+                    }
+                }
+            }
+        `;
+
+        const gqlResult = await testHelper.executeGraphQL(mutation);
+
+        expect(gqlResult.errors).toBeFalsy();
+        expect(gqlResult.data).toEqual({
+            [Movie.operations.create]: {
+                [Movie.plural]: expect.toIncludeSameMembers([
+                    { duration: duration1.toString() },
+                    { duration: duration2.toString() },
+                ]),
+            },
+        });
+    });
+});
