+++ packages/graphql/tests/api-v6/integration/mutations/create/types/localTime.int.test.ts
@@ -18,10 +18,10 @@
  */
 
 import neo4jDriver from "neo4j-driver";
-import type { UniqueType } from "../../../../utils/graphql-types";
-import { TestHelper } from "../../../../utils/tests-helper";
+import type { UniqueType } from "../../../../../utils/graphql-types";
+import { TestHelper } from "../../../../../utils/tests-helper";
 
-describe("Create Nodes with Date fields", () => {
+describe("Create Nodes with LocalTime fields", () => {
     const testHelper = new TestHelper({ v6Api: true });
     let Movie: UniqueType;
 
@@ -29,7 +29,7 @@ describe("Create Nodes with Date fields", () => {
         Movie = testHelper.createUniqueType("Movie");
         const typeDefs = /* GraphQL */ `
         type ${Movie.name} @node {
-            date: Date
+            localTime: LocalTime
         }
     `;
         await testHelper.initNeo4jGraphQL({ typeDefs });
@@ -39,22 +39,23 @@ describe("Create Nodes with Date fields", () => {
         await testHelper.close();
     });
 
-    test("should be able to create nodes with date fields", async () => {
-        const date1 = new Date(1716904582368);
-        const date2 = new Date(1736900000000);
-        const neoDate1 = neo4jDriver.types.Date.fromStandardDate(date1);
-        const neoDate2 = neo4jDriver.types.Date.fromStandardDate(date2);
+    test("should be able to create nodes with LocalTime fields", async () => {
+        const time1 = new Date("2024-02-17T11:49:48.322Z");
+        const time2 = new Date("2025-02-17T12:49:48.322Z");
+
+        const neoTime1 = neo4jDriver.LocalTime.fromStandardDate(time1);
+        const neoTime2 = neo4jDriver.LocalTime.fromStandardDate(time2);
 
         const mutation = /* GraphQL */ `
             mutation {
                 ${Movie.operations.create}(input: [
-                        { node: { date: "${neoDate1.toString()}" } }
-                        { node: { date: "${neoDate2.toString()}" } }
+                        { node: { localTime: "${neoTime1.toString()}" } }
+                        { node: { localTime: "${neoTime2.toString()}" } }
                     ]) {
-                        ${Movie.plural} {
-                            date
-                        }
+                    ${Movie.plural} {
+                        localTime
                     }
+                }
             }
         `;
 
@@ -64,8 +65,8 @@ describe("Create Nodes with Date fields", () => {
         expect(gqlResult.data).toEqual({
             [Movie.operations.create]: {
                 [Movie.plural]: expect.toIncludeSameMembers([
-                    { date: neoDate1.toString() },
-                    { date: neoDate2.toString() },
+                    { localTime: neoTime1.toString() },
+                    { localTime: neoTime2.toString() },
                 ]),
             },
         });
